+++ test/unit/enum.test.js
@@ -3,12 +3,35 @@
 const fs = require('fs').promises
 const path = require('path')
 const cds2ts = require('../../lib/compile')
-const { ASTWrapper, check, JSASTWrapper } = require('../ast')
+const { ASTWrapper, check, JSASTWrapper, checkFunction } = require('../ast')
 const { locations } = require('../util')
 
 const dir = locations.testOutput('enums_test')
 
 // FIXME: missing: inline enums (entity Foo { bar: String enum { ... }})
+describe('Enum Action Parameters', () => {
+    let astw
+
+    beforeEach(async () => await fs.unlink(dir).catch(() => {}))
+    beforeAll(async () => {
+        const paths = await cds2ts
+            .compileFromFile(locations.unit.files('enums/actions.cds'), { outputDirectory: dir, inlineDeclarations: 'structured' })
+        astw = new ASTWrapper(path.join(paths[1], 'index.ts'))
+    })
+    
+    test('Coalescing Assignment Present', () => {
+        const actions = astw.getAspectProperty('_FoobarAspect', 'actions')
+        checkFunction(actions.type.members.find(fn => fn.name === 'f'), {
+            parameterCheck: ({members: [fst]}) => fst.name === 'p'
+                && check.isUnionType(fst.type, [
+                    t => check.isLiteral(t, 'A'),
+                    t => check.isLiteral(t, 'b'),
+                ])
+        })
+    }) 
+})
+
+
 describe('Nested Enums', () => {
     let astw
 
@@ -26,8 +49,6 @@ describe('Nested Enums', () => {
         const { left } = enm.expression
         // not checking the entire object chain here...
         expect(left.property.name).toBe('someEnumProperty')
-
-        console.log(42)
     }) 
 })